Save loads of memory (huges zlib buffers) by disabling SSL compression.
The OpenSSL compression's 500KB+ per connection penalty doesn't fit well
with the high concurrency and high scalability capabilities we love in stud.

SSL_OP_NO_COMPRESSION is only available since openssl 1.0.0+.

Nodejs has a pragmatic hack to disable compression even on 0.9.8,
I didn't dare to suggest that for stud, but I would happily to port
and test that for stud if you ask ;).
http://journal.paul.querna.org/articles/2011/04/05/openssl-memory-use/
https://github.com/joyent/node/blob/master/src/node_crypto.cc#L4011

Benchmark for the patch available here :
http://bpineau.livejournal.com/1663.html