Expand man page make target argument.
It appears that on FreeBSD systems the $< variable doesn't work
as expected, and is evaluated to an empty string.

Since this is pretty simple stuff, and we don't have or will have a
whole lot of them, I'm taking the easy route and expanding the
variable.

Reported by:	Ryan Steinmetz