Fix intermitten CI build errors.
Instead of starting, requiring a successful DNS lookup and then running
a request, just parse the old configuration files and check for syntax
errors.

Related commit: d73bf34

Not as complete as before, but it is more important to be able to rely
on the test suite as a whole. It should also take a bit less time to
run, which is always a plus for a living test suite.

We can revisit this if we ever put a TCP server in here, that can reply
without having to traverse the unreliable internet.