Manual page for hitch.conf
The information from hitch.conf.example is made into a man page.

There is now a reference in the hitch man file to the hitch.conf
manual.

Thanks to Lasse Karstensen for helpful feedback.

The hitch.conf.example file remains, but should either be generated
from the hitch.conf man file, or removed completely, as the
information is too spread out.