Simplify example configuration to bare minimum.
If default.vcl in Varnish have taught us anything, it is that
configurations stanzas live for a long long time when they are on a
working installation/rig.

Try to avoid having these defaults being unchangeable in the future
by not including them here, and rather have them as proper Hitch
defaults in configuration.c.

It has the added benefit of making the file shorter and more easily
understood by the user.