Set default cipher list to current best practice.
As far as I can tell, this is the current best practice. The standard
test sites online give this cipher list good grades.

This change was discussed in length internally.

I believe it is more important to be safe by default than to perhaps
ship the wrong default for a while.

The reason for this is that if users put the cipher list into their
configuration (like they had to, up until now), the majority will not
change it if Hitch is upgraded.

In the case of a broken cipher or cipher mode we won't be able to reach
all of them, and they will run an insecure configuration "forever".

If (more likely: when) we change the default some time in the future, we
implicitly reach them next time they upgrade their Hitch version.

Users should still definitively do their own research and pay attention,
but unless they have special requirements Hitch should need as little
configuration as possible.