Add a default OCSP stapling cache directory.
Use /var/lib/hitch/ for this by default.

I'm not entirely happy to put all of these paths into the
source itself, but we really want this to be default on and
as simple as possible to use.

Packagers should make /var/lib/hitch writeable by the user
hitch will be running as.

People running from source should mkdir and chown this after
they have run make install. This is not done automatically.