Make -Wno-strict-aliasing a flag only for libev
Because of macro constructs leading to strict aliasing warnings, it was
disabled for the whole code base, now it's injected in $(EV_CFLAGS) to
keep it contained to binaries linking to libev. Which is still most of
the code base but no longer all of it.

The default hitch warnings can't be part of the CFLAGS, they would
otherwise enable the warning again. Now we have a clear distinction
between flags we want for the code base (HITCH_CFLAGS), flags we inherit
from dependencies (*_FLAGS) and flags that may be set by the user to
override them (CFLAGS).