Avoid leaking a zombie process for the OCSP child
Turns out we mistakenly attempted to restart the process twice on a HUP.
First in the reconfigure code, and then again in do_wait.

On the second time around, the PID of the initial process is since
overwritten, and it will be leaked as a zombie proc.

Thanks to github user @dward for his analysis of the problem.

Fixes: #167