mv common.sh hitch_test.sh
The idea is to have a less generic name for the test fixtures, and one
that can be sourced as-is also. For that the $TESTDIR is added to the
PATH.

We don't need to build absolute paths from the top of the source tree
since when referencing subdirectories, so a little polish went there
too.