Wrap openssl s_client commands in a function
It takes care of feeding them newlines or checking that options like
-alpn are supported. Start using it in NPN and ALPN test cases.

About this trick:

    ! grep ERROR $DUMPFILE

Instead of using run_cmd we can grep for errors, but this will fail when
there are no errors. With `set -e` enforced in tests 15 and 16 the lack
of error would fail the whole test. So instead the exit status must be
non-zero (thanks to the ! prefix) and when errors occur they are printed
thanks to the lack of a -q option.