Rework the handling of optional dependencies
There are multiple changes here, first we always ship the documentation
in the dist archive. This way, building a release from a dist archive no
longer requires rst2man. Unless downstream distributors modify the RST
files they don't need to rebuild the manuals.

The same is done for lex and yacc (or flex and bison) because we already
ship cfg_lex.c and cfg_parser.c in the dist archive. Unless downstream
distributors modify cfg_lex.l or cfg_parser.y they don't need to rebuild
those files.

Fixes #263

Instead of calling the individual autotools, the bootstrap script now
handles it all at once with autoreconf.

Fixes #252
Closes #306

The last significant change is that ./bootstrap now invokes ./configure
and may take command line arguments such as --enable-sessioncache to
pass them on to the configure script:

    ./bootstrap --enable-sessioncache

Since the bootstrap script is meant for Hitch development, and not
redistribution it requires rst2man, lex/flex, yacc/bison to be present.

This means that we will need to adjust or development habits (this is
similar to what ./autogen.des does in Varnish) and build infrastrucure.