Centralize PROXY v2 definitions
This declares anything relevant from the spec, and more, to get a
central location within reach to avoid fetching the spec anytime it
comes up.

This also gets rid of a bunch of magic numbers whenever PROXY v2 is
involved.

Also, a consistent PP2_ (or pp2_) namespace is used for the header
definitions, even though the spec uses a mix of PP2_, pp2_ and
proxy_ throughout.

Please note a couple of XXX comments added while I was reviewing
this part of the code.