Don't rely on the OpenSSL version to detect ALPN
A vendor may back-port the feature to an older release, so we might as
well look for the presence of one of its basic functions. This includes
NPN since the code uses them in tandem, and keeps the ability to exclude
NPN with the OPENSSL_NO_NEXTPROTONEG macro since OpenSSL itself relies
on that internally.

Except that currently Hitch won't build when OPENSSL_NO_NEXTPROTONEG is
defined, which was already the case before this change.