Wrap the memcpy+memset padding pattern in a macro
Semi automated with the following Coccinelle patch:

    @@
    expression dst, dlen, src, slen;
    @@

    - memcpy(dst, src, slen);
    - if (slen < dlen)
    -       memset(dst + slen, 0, dlen - slen);
    + shsess_memcpypad(dst, dlen, src, slen);

There are other cases where the if statement encompasses both memcpy and
memset operations. However the probability that the key doesn't require
padding (and is exactly the maximum length long) is low, so there's no
point in adding a branch to save a very likely copy operation.

Slightly better diff with the --ignore-all-space option.