Stabilization of the ticket resumption test
As @l0stman pointed out on GitHub TLS 1.3 uses 0-RTT instead of session
resumption, disabling TLS 1.3 for this test makes it pass consistently.

However, when openssl s_client is run with strace the test passes and
the log shows that a TLS 1.3 session was indeed reused by the curl
client. This test would also sometimes pass even without the help of
strace.

Refs #292