Do the socket bind & listen in the child worker process
This addresses the issue with imbalanced load distribution between the
worker processes.

In the change introduced the master process will still do its initial
bind & listen as a test to see if we're good to go, after which it
will close its bound sockets and let the workers create their own
sockets and bind individually.

Fixes: #142